feat: split sigmastudio protocol handler common parts into a base module, add TCPIP1701 support
